+++ src/Quests/RewardType.cs
@@ -1,0 +1,9 @@
+﻿namespace QuestFramework.Quests
+{
+    public enum RewardType
+    {
+        Money = 2,
+        Object = 4,
+        Weapon = 8,
+    }
+}