+++ src/Structures/QuestLogColors.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace QuestFramework.Structures
+{
+    public class QuestLogColors
+    {
+        public int TitleColor { get; set; } = -1;
+        public int TextColor { get; set; } = -1;
+        public int ObjectiveColor { get; set; } = -1;
+    }
+}